requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.5.0
    # via
    #   starlette
    #   watchgod
asgiref==3.5.0
    # via uvicorn
atomicwrites==1.4.0
    # via pytest
attrs==21.4.0
    # via pytest
bcrypt==3.2.0
    # via -r requirements.in
cffi==1.15.0
    # via bcrypt
click==8.0.4
    # via uvicorn
colorama==0.4.4
    # via
    #   click
    #   pytest
    #   uvicorn
ecdsa==0.17.0
    # via python-jose
fastapi==0.75.0
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
h11==0.13.0
    # via uvicorn
httptools==0.4.0
    # via uvicorn
idna==3.3
    # via anyio
iniconfig==1.1.1
    # via pytest
mypy==0.941
    # via -r requirements.in
mypy-extensions==0.4.3
    # via mypy
packaging==21.3
    # via pytest
passlib==1.7.4
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==1.9.0
    # via fastapi
pyparsing==3.0.7
    # via packaging
pytest==7.1.1
    # via -r requirements.in
python-dotenv==0.19.2
    # via
    #   -r requirements.in
    #   uvicorn
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
pyyaml==6.0
    # via uvicorn
rsa==4.8
    # via python-jose
six==1.16.0
    # via
    #   bcrypt
    #   ecdsa
    #   python-multipart
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.32
    # via -r requirements.in
starlette==0.17.1
    # via fastapi
tomli==2.0.1
    # via
    #   mypy
    #   pytest
typing-extensions==4.1.1
    # via
    #   mypy
    #   pydantic
uvicorn[standard]==0.17.6
    # via -r requirements.in
watchgod==0.8
    # via uvicorn
websockets==10.2
    # via uvicorn
